requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
pillow>=9.0.0
numpy>=1.24.0
scipy>=1.10.0

# CLIP model
open_clip_torch>=2.24.0

# Data handling
pycocotools>=2.0.6
requests>=2.28.0
tqdm>=4.65.0

# UI
gradio>=4.0.0

# Translation for Vietnamese support
deep-translator>=1.11.0

# Utilities
matplotlib>=3.7.0
scikit-learn>=1.3.0
pandas>=2.0.0
seaborn>=0.12.0
